requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.4
beautifulsoup4==4.9.3
billiard==3.6.4.0
black==21.5b1
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.3
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.0
kombu==5.0.2
lxml==4.6.3
MarkupSafe==2.0.0
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
prompt-toolkit==3.0.18
pyparsing==2.4.7
pytz==2021.1
regex==2021.4.4
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
